Move wakeup source impls
